fix: remove notify support from erofs
